👽 Remove deprecated option for `grep`
### Changes Description

On latest version of `grep`, it stop supporting the `GREP_OPTIONS`
environment variable in favor of using aliases.

From Grep's manual:

> `GREP_OPTIONS`
  This variable specifies default options to be placed in front of any
  explicit options. As this causes problems when writing portable
  scripts, this feature will be removed in a future release of grep, and
  grep warns if it is used. Please use an alias or script instead.

This commit removes the deprecated option and add an alias with the same
functionality.

### Testing

Use `grep` in a text file and make sure that matches are highlighted in
different color.

## Related items

- See also: [Grep Environment Variables](https://www.gnu.org/software/grep/manual/html_node/Environment-Variables.html)