🐛 Fix _has function and add _has_alias
## Changes Description
The `_has` function was not preventing error outputs from being
returned. This commit changes this behavior and filter all outputs from
the `whence` command (e.g. `>/dev/null 2>&1`).

This commit also introduces a new base function called `_has_alias`. It
works exactly like the `_has` function, but it checks whether an alias
is defined or not.